Add support for typed queries
Add the graphql-codegen package to add support for typed queries
