+++ app/routes/app._index.tsx
@@ -61,7 +61,7 @@ export const action = async ({ request }: ActionFunctionArgs) => {
   const responseJson = await response.json();
 
   return json({
-    product: responseJson.data.productCreate.product,
+    product: responseJson.data?.productCreate?.product,
   });
 };
