+++ package.json
@@ -16,7 +16,8 @@
     "setup": "prisma generate && prisma migrate deploy",
     "lint": "eslint --cache --cache-location ./node_modules/.cache/eslint .",
     "shopify": "shopify",
-    "prisma": "prisma"
+    "prisma": "prisma",
+    "graphql-codegen": "graphql-codegen"
   },
   "dependencies": {
     "@prisma/client": "^4.13.0",
@@ -27,9 +28,9 @@
     "@shopify/app": "^3.50.0",
     "@shopify/cli": "^3.50.0",
     "@shopify/polaris": "^12.0.0",
-    "@shopify/shopify-api": "^8.0.2",
-    "@shopify/shopify-app-remix": "^2.1.0",
-    "@shopify/shopify-app-session-storage-prisma": "^2.0.0",
+    "@shopify/shopify-api": "^9.0.1",
+    "@shopify/shopify-app-remix": "^2.3.0",
+    "@shopify/shopify-app-session-storage-prisma": "^2.0.3",
     "isbot": "^4.1.0",
     "prisma": "^4.13.0",
     "react": "^18.2.0",
@@ -37,6 +38,7 @@
   },
   "devDependencies": {
     "@remix-run/eslint-config": "^2.0.0",
+    "@shopify/api-codegen-preset": "^0.0.1",
     "@shopify/app-bridge-types": "^0.0.3",
     "@types/eslint": "^8.40.0",
     "@types/node": "^20.6.3",
@@ -52,5 +54,7 @@
       "extensions/*"
     ]
   },
-  "trustedDependencies": ["@shopify/plugin-cloudflare"]
+  "trustedDependencies": [
+    "@shopify/plugin-cloudflare"
+  ]
 }
